requirements.txt
streamlit==1.32.2
tensorflow-cpu==2.15.0
keras==2.15.0
numpy<2.0.0
